[FrameworkBundle][HttpKernel] Configure ErrorHandler on boot
