+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/debug_prod.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
+use Symfony\Component\HttpKernel\Debug\ErrorHandlerConfigurator;
 use Symfony\Component\HttpKernel\Debug\FileLinkFormatter;
 use Symfony\Component\HttpKernel\EventListener\DebugHandlersListener;
 
@@ -18,9 +19,9 @@ return static function (ContainerConfigurator $container) {
     $container->parameters()->set('debug.error_handler.throw_at', -1);
 
     $container->services()
-        ->set('debug.debug_handlers_listener', DebugHandlersListener::class)
+        ->set('debug.error_handler_configurator', ErrorHandlerConfigurator::class)
+            ->public()
             ->args([
-                null, // Exception handler
                 service('monolog.logger.php')->nullOnInvalid(),
                 null, // Log levels map for enabled error levels
                 param('debug.error_handler.throw_at'),
@@ -28,8 +29,10 @@ return static function (ContainerConfigurator $container) {
                 param('kernel.debug'),
                 service('monolog.logger.deprecation')->nullOnInvalid(),
             ])
-            ->tag('kernel.event_subscriber')
             ->tag('monolog.logger', ['channel' => 'php'])
+
+        ->set('debug.debug_handlers_listener', DebugHandlersListener::class)
+            ->tag('kernel.event_subscriber')
 
         ->set('debug.file_link_formatter', FileLinkFormatter::class)
             ->args([param('debug.file_link_format')])
