+++ src/Symfony/Component/HttpKernel/EventListener/DebugHandlersListener.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\HttpKernel\EventListener;
 
-use Psr\Log\LoggerInterface;
 use Symfony\Component\Console\ConsoleEvents;
 use Symfony\Component\Console\Event\ConsoleEvent;
 use Symfony\Component\Console\Output\ConsoleOutputInterface;
@@ -21,7 +20,7 @@ use Symfony\Component\HttpKernel\Event\KernelEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
 
 /**
- * Configures errors and exceptions handlers.
+ * Sets an exception handler.
  *
  * @author Nicolas Grekas <p@tchwork.com>
  *
@@ -33,35 +32,19 @@ class DebugHandlersListener implements EventSubscriberInterface
 {
     private string|object|null $earlyHandler;
     private ?\Closure $exceptionHandler;
-    private ?LoggerInterface $logger;
-    private ?LoggerInterface $deprecationLogger;
-    private array|int|null $levels;
-    private ?int $throwAt;
-    private bool $scream;
-    private bool $scope;
     private bool $firstCall = true;
     private bool $hasTerminatedWithException = false;
 
     /**
-     * @param callable|null  $exceptionHandler A handler that must support \Throwable instances that will be called on Exception
-     * @param array|int|null $levels           An array map of E_* to LogLevel::* or an integer bit field of E_* constants
-     * @param int|null       $throwAt          Thrown errors in a bit field of E_* constants, or null to keep the current value
-     * @param bool           $scream           Enables/disables screaming mode, where even silenced errors are logged
-     * @param bool           $scope            Enables/disables scoping mode
+     * @param callable|null $exceptionHandler A handler that must support \Throwable instances that will be called on Exception
      */
-    public function __construct(callable $exceptionHandler = null, LoggerInterface $logger = null, array|int|null $levels = \E_ALL, ?int $throwAt = \E_ALL, bool $scream = true, bool $scope = true, LoggerInterface $deprecationLogger = null)
+    public function __construct(callable $exceptionHandler = null)
     {
         $handler = set_exception_handler('is_int');
         $this->earlyHandler = \is_array($handler) ? $handler[0] : null;
         restore_exception_handler();
 
         $this->exceptionHandler = null === $exceptionHandler ? null : $exceptionHandler(...);
-        $this->logger = $logger;
-        $this->levels = $levels ?? \E_ALL;
-        $this->throwAt = \is_int($throwAt) ? $throwAt : (null === $throwAt ? null : ($throwAt ? \E_ALL : null));
-        $this->scream = $scream;
-        $this->scope = $scope;
-        $this->deprecationLogger = $deprecationLogger;
     }
 
     /**
@@ -77,40 +60,6 @@ class DebugHandlersListener implements EventSubscriberInterface
         }
         $this->firstCall = $this->hasTerminatedWithException = false;
 
-        $handler = set_exception_handler('is_int');
-        $handler = \is_array($handler) ? $handler[0] : null;
-        restore_exception_handler();
-
-        if (!$handler instanceof ErrorHandler) {
-            $handler = $this->earlyHandler;
-        }
-
-        if ($handler instanceof ErrorHandler) {
-            if ($this->logger || $this->deprecationLogger) {
-                $this->setDefaultLoggers($handler);
-                if (\is_array($this->levels)) {
-                    $levels = 0;
-                    foreach ($this->levels as $type => $log) {
-                        $levels |= $type;
-                    }
-                } else {
-                    $levels = $this->levels;
-                }
-
-                if ($this->scream) {
-                    $handler->screamAt($levels);
-                }
-                if ($this->scope) {
-                    $handler->scopeAt($levels & ~\E_USER_DEPRECATED & ~\E_DEPRECATED);
-                } else {
-                    $handler->scopeAt(0, true);
-                }
-                $this->logger = $this->deprecationLogger = $this->levels = null;
-            }
-            if (null !== $this->throwAt) {
-                $handler->throwAt($this->throwAt, true);
-            }
-        }
         if (!$this->exceptionHandler) {
             if ($event instanceof KernelEvent) {
                 if (method_exists($kernel = $event->getKernel(), 'terminateWithException')) {
@@ -136,38 +85,18 @@ class DebugHandlersListener implements EventSubscriberInterface
             }
         }
         if ($this->exceptionHandler) {
+            $handler = set_exception_handler('is_int');
+            $handler = \is_array($handler) ? $handler[0] : null;
+            restore_exception_handler();
+
+            if (!$handler instanceof ErrorHandler) {
+                $handler = $this->earlyHandler;
+            }
+
             if ($handler instanceof ErrorHandler) {
                 $handler->setExceptionHandler($this->exceptionHandler);
             }
             $this->exceptionHandler = null;
-        }
-    }
-
-    private function setDefaultLoggers(ErrorHandler $handler): void
-    {
-        if (\is_array($this->levels)) {
-            $levelsDeprecatedOnly = [];
-            $levelsWithoutDeprecated = [];
-            foreach ($this->levels as $type => $log) {
-                if (\E_DEPRECATED == $type || \E_USER_DEPRECATED == $type) {
-                    $levelsDeprecatedOnly[$type] = $log;
-                } else {
-                    $levelsWithoutDeprecated[$type] = $log;
-                }
-            }
-        } else {
-            $levelsDeprecatedOnly = $this->levels & (\E_DEPRECATED | \E_USER_DEPRECATED);
-            $levelsWithoutDeprecated = $this->levels & ~\E_DEPRECATED & ~\E_USER_DEPRECATED;
-        }
-
-        $defaultLoggerLevels = $this->levels;
-        if ($this->deprecationLogger && $levelsDeprecatedOnly) {
-            $handler->setDefaultLogger($this->deprecationLogger, $levelsDeprecatedOnly);
-            $defaultLoggerLevels = $levelsWithoutDeprecated;
-        }
-
-        if ($this->logger && $defaultLoggerLevels) {
-            $handler->setDefaultLogger($this->logger, $defaultLoggerLevels);
         }
     }
